Move root crate re-export docs to the root.
