+++ druid/src/lens/lens.rs
@@ -12,51 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Support for lenses, a way of focusing on subfields of data.
-//!
-//! Lenses are useful whenever a widget only needs access to a subfield of a larger struct or
-//! generally access to part of a larger value.
-//!
-//! For example: If one wants to embed a [`TextBox`] in a widget with a `Data` type
-//! that is not `String`, they need to specify how to access a `String` from within the `Data`.
-//!
-//! [`TextBox`]: ../widget/struct.TextBox.html
-//! ```
-//! use druid::{Data, Lens, Widget, WidgetExt, widget::{TextBox, Flex}};
-//!
-//! #[derive(Clone, Debug, Data, Lens)]
-//! struct MyState {
-//!     search_term: String,
-//!     scale: f64,
-//!     // ...
-//! }
-//!
-//!
-//! fn my_sidebar() -> impl Widget<MyState> {
-//!     // `TextBox` is of type `Widget<String>`
-//!     // via `.lens` we get it to be of type `Widget<MyState>`.
-//!     // `MyState::search_term` is a lens generated by the `derive(Lens)` macro,
-//!     // that provides access to the search_term field.
-//!     let searchbar = TextBox::new().lens(MyState::search_term);
-//!
-//!     // ...
-//!
-//!     // We can now use `searchbar` just like any other `Widget<MyState>`
-//!     Flex::column().with_child(searchbar)
-//! }
-//! ```
-
 use std::marker::PhantomData;
 use std::ops;
 use std::sync::Arc;
 
-pub use druid_derive::Lens;
-
 use crate::kurbo::Size;
-use crate::{
-    BoxConstraints, Data, Env, Event, EventCtx, LayoutCtx, LifeCycle, LifeCycleCtx, PaintCtx,
-    UpdateCtx, Widget, WidgetId,
-};
+use crate::widget::prelude::*;
+use crate::Data;
 
 /// A lens is a datatype that gives access to a part of a larger
 /// data structure.
